Partial fix of msg issues, restructuring
