+++ igvc_platform/src/tests/CMakeLists.txt
@@ -5,3 +5,8 @@ target_link_libraries(TestJoystickDriver ${catkin_LIBRARIES})
 add_rostest_gtest(TestDifferentialDrive test/test_differential_drive.test test_differential_drive.cpp)
 add_dependencies(TestDifferentialDrive ${catkin_EXPORTED_TARGETS} differential_drive)
 target_link_libraries(TestDifferentialDrive ${catkin_LIBRARIES})
+
+add_rostest_gtest(TestScanToPointCloud test/test_scan_to_pointcloud.test test_scan_to_pointcloud.cpp)
+add_dependencies(TestScanToPointCloud ${catkin_EXPORTED_TARGETS} scan_to_pointcloud)
+target_link_libraries(TestScanToPointCloud ${catkin_LIBRARIES})
+
