+++ igvc_platform/src/tests/test/test_scan_to_pointcloud.test
@@ -1,0 +1,4 @@
+<launch>
+    <node name="scan_to_pointcloud" pkg="igvc_gazebo" type="scan_to_pointcloud" output="screen" required="true"/>
+    <test test-name="test_scan_to_pointcloud" pkg="igvc_platform" type="TestScanToPointCloud"/>
+</launch>
